rename root module
