+++ src/main/java/com/example/soundtracks/SoundtracksApplication.java
@@ -1,4 +1,4 @@
-package com.example.spotifydemo;
+package com.example.soundtracks;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
