+++ src/main/java/com/example/soundtracks/WebConfiguration.java
@@ -1,4 +1,4 @@
-package com.example.spotifydemo;
+package com.example.soundtracks;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
